+++ CHANGELOG.md
@@ -27,7 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added changelog.
 - Changed way how new libraries are created.
 
-## [1.0.2] - 
+## [1.1.0] - 
 
 ### Added
 
@@ -36,7 +36,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added option to automatically create a backup when saving files.
 - Added command for renaming a series.
 - Added report to print list of comics.
+- Added fields to enter purchase price and estimated value.
+- Added global currency.
 
 ### Fixed
 
 - Fixed export of images.
+- Fixed deleting of optional values (issue number, year) while editing.